bora la 2

+++ Proj_2/Entity.cpp
@@ -12,7 +12,7 @@ Vector3* Entity::getPosition(){
     return _position;
 }
 
-Vector3* Entity::setPosition(double x, double y, double z){
+Vector3* Entity::setPosition(float x, float y, float z){
     _position->set(x, y, z);
     return _position;
 }
