+++ Proj_2/Entity.h
@@ -14,7 +14,7 @@ public:
     Entity();
     ~Entity();
     Vector3* getPosition();
-    Vector3* setPosition(double x, double y, double z);
+    Vector3* setPosition(float x, float y, float z);
     Vector3* setPosition(Vector3 &p);
 };
